+++ src/api/qualicharge/migrations/env.py
@@ -8,6 +8,16 @@ from sqlmodel import SQLModel
 
 from qualicharge.conf import settings
 
+# Nota bene: be sure to import all models that need to be migrated here
+from qualicharge.schemas.static import (  # noqa: F401
+    Amenageur,
+    Enseigne,
+    Localisation,
+    Operateur,
+    PointDeCharge,
+    Station,
+)
+
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
 config = context.config
@@ -70,7 +80,10 @@ def run_migrations_online() -> None:
     )
 
     with connectable.connect() as connection:
-        context.configure(connection=connection, target_metadata=target_metadata)
+        context.configure(
+            connection=connection,
+            target_metadata=target_metadata,
+        )
 
         with context.begin_transaction():
             context.run_migrations()